requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
jinja2==3.1.5
python-multipart==0.0.9
sqlalchemy==2.0.36
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
qrcode==7.4.2
pillow==10.4.0
itsdangerous==2.2.0
